Rename StudentProfile routes. Update create routes
